app/requirements.txt
flask
python-dotenv
pandas
authlib
requests
Werkzeug==2.2.2
sqlalchemy
sentry_sdk
pymysql